+++ src/poller.h
@@ -33,42 +33,22 @@ public:
         poll->data = this;
         err = uv_poll_init_socket(loop, poll, fd);
         if (err != 0) return err;
-        uv_unref(poll);
 
         readable_timer->data = this;
         err = uv_timer_init(loop, readable_timer);
         if (err != 0) return err;
-        uv_unref(readable_timer);
 
         writable_timer->data = this;
         err = uv_timer_init(loop, writable_timer);
         if (err != 0) return err;
-        uv_unref(writable_timer);
 
         finalize = finalizer;
         return 0;
     }
 
-    inline void Ref() {
-        /* Avoid segfault if accidentally called after closing. */
-        if (poll != nullptr) {
-            uv_ref(poll);
-        }
-    }
-
-    inline void Unref() {
-        /* Avoid segfault if accidentally called after closing. */
-        if (poll != nullptr) {
-            uv_unref(poll);
-        }
-    }
-
     /* Safely close and release all handles. This can be called before
        destruction to release resources early. */
-    inline void Close(bool cancel = false) {
-        /* Cancels watched events, don't trigger. */
-        if (cancel) events = 0;
-
+    inline void Close() {
         /* Trigger all watched events manually, which causes any pending
            operation to succeed or fail immediately. */
         if (events) Trigger(events);