+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Unreleased
+
+* Reading from event observers now prevent the Node process from exiting, even if the underlying socket is no longer being used.
+
+* Reading or writing on unbound and unconnected sockets now prevent the Node process from exiting.
+
 ### v6.0.0-beta.2
 
 * Fix omission of "zeromq/v5-compat" export.